Remove devise authorization
Supporting getting a user name from a rails cookie is difficult, and
also requires integration with the rails application database to look up
users in the Users table, which is fragile and could change with every
application. The preferred way to do an integration with a
rails application is to use nginx and X-Accel-Redirect.

Added documentation to auth package stating this. Perhaps also add the
nginx configurations?

As a side-effect we can remove a lot of the vendored dependencies.